+++ requirements.txt
@@ -1,1 +1,1 @@
-ansible==2.9.10
+ansible==2.9.9
